Rearrange modules
