+++ crates/relayer-all-in-one/src/all_for_one/relay.rs
@@ -9,7 +9,6 @@ use ibc_relayer_components::relay::traits::connection::open_handshake::CanRelayC
 use ibc_relayer_components::relay::traits::connection::open_init::CanInitConnection;
 use ibc_relayer_components::relay::traits::event_relayer::CanRelayEvent;
 use ibc_relayer_components::relay::traits::ibc_message_sender::{CanSendIbcMessages, MainSink};
-use ibc_relayer_components::relay::traits::messages::update_client::CanBuildUpdateClientMessage;
 use ibc_relayer_components::relay::traits::packet::HasRelayPacket;
 use ibc_relayer_components::relay::traits::packet_clear::CanClearReceivePackets;
 use ibc_relayer_components::relay::traits::packet_filter::CanFilterPackets;
@@ -18,6 +17,7 @@ use ibc_relayer_components::relay::traits::packet_relayers::ack_packet::CanRelay
 use ibc_relayer_components::relay::traits::packet_relayers::receive_packet::CanRelayReceivePacket;
 use ibc_relayer_components::relay::traits::packet_relayers::timeout_unordered_packet::CanRelayTimeoutUnorderedPacket;
 use ibc_relayer_components::relay::traits::target::{DestinationTarget, SourceTarget};
+use ibc_relayer_components::relay::traits::update_client::CanBuildUpdateClientMessage;
 use ibc_relayer_components_extra::relay::components::packet_relayers::retry::SupportsPacketRetry;
 
 use crate::all_for_one::chain::AfoChain;
