+++ crates/relayer-components/src/relay/components/mod.rs
@@ -1,5 +1,6 @@
 pub mod auto_relayers;
 pub mod event_relayers;
 pub mod message_senders;
+pub mod packet_filters;
 pub mod packet_relayers;
 pub mod update_client;
