+++ crates/relayer-components/src/relay/components/packet_relayers/timeout_unordered/mod.rs
@@ -1,2 +1,1 @@
 pub mod timeout_unordered_packet;
-pub mod wait_timeout;
