+++ crates/relayer-components/src/relay/components/update_client/wait.rs
@@ -7,8 +7,8 @@ use crate::chain::traits::wait::CanWaitChainReachHeight;
 use crate::logger::traits::level::HasBaseLogLevels;
 use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::logs::logger::CanLogRelayTarget;
-use crate::relay::traits::messages::update_client::UpdateClientMessageBuilder;
 use crate::relay::traits::target::ChainTarget;
+use crate::relay::traits::update_client::UpdateClientMessageBuilder;
 use crate::runtime::traits::runtime::HasRuntime;
 use crate::std_prelude::*;
