+++ crates/relayer-components/src/relay/impls/connection/open_ack.rs
@@ -5,11 +5,11 @@ use crate::chain::traits::message_builders::connection::{
     CanBuildConnectionHandshakeMessages, CanBuildConnectionHandshakePayloads,
 };
 use crate::chain::traits::queries::status::CanQueryChainHeight;
-use crate::relay::impls::update_client::CanSendUpdateClientMessage;
 use crate::relay::traits::chains::HasRelayChains;
 use crate::relay::traits::connection::open_ack::ConnectionOpenAckRelayer;
 use crate::relay::traits::ibc_message_sender::{CanSendSingleIbcMessage, MainSink};
 use crate::relay::traits::target::{DestinationTarget, SourceTarget};
+use crate::relay::traits::update_client::CanSendUpdateClientMessage;
 use crate::std_prelude::*;
 
 /**
