+++ crates/relayer-components/src/relay/impls/mod.rs
@@ -2,5 +2,3 @@ pub mod channel;
 pub mod client;
 pub mod connection;
 pub mod packet_clear;
-pub mod packet_filters;
-pub mod update_client;
