--- crates/relayer-components/src/relay/traits/messages/mod.rs
@@ -1,4 +1,0 @@
-pub mod timeout_channel_closed_packet;
-pub mod timeout_ordered;
-pub mod timeout_unordered_packet;
-pub mod update_client;
