+++ crates/relayer-components/src/relay/traits/mod.rs
@@ -6,7 +6,6 @@ pub mod connection;
 pub mod event_relayer;
 pub mod ibc_message_sender;
 pub mod logs;
-pub mod messages;
 pub mod packet;
 pub mod packet_clear;
 pub mod packet_filter;
@@ -14,3 +13,4 @@ pub mod packet_relayer;
 pub mod packet_relayers;
 pub mod target;
 pub mod two_way;
+pub mod update_client;
