+++ crates/relayer-mock/src/relayer_mock/base/impls/relay.rs
@@ -5,9 +5,9 @@ use ibc_relayer_components::core::traits::component::DelegateComponent;
 use ibc_relayer_components::core::traits::error::HasErrorType;
 use ibc_relayer_components::logger::traits::has_logger::{HasLogger, HasLoggerType};
 use ibc_relayer_components::relay::traits::chains::HasRelayChains;
-use ibc_relayer_components::relay::traits::messages::update_client::UpdateClientMessageBuilder;
 use ibc_relayer_components::relay::traits::packet_relayers::lock::HasPacketLock;
 use ibc_relayer_components::relay::traits::target::{DestinationTarget, SourceTarget};
+use ibc_relayer_components::relay::traits::update_client::UpdateClientMessageBuilder;
 use ibc_relayer_components::runtime::traits::runtime::HasRuntime;
 use ibc_relayer_runtime::tokio::logger::tracing::TracingLogger;
 use std::vec;
